+++ registry/components/ui/field/index.ts
@@ -1,0 +1,5 @@
+export { default as Field } from "./field.svelte";
+export { default as FieldLabel } from "./field-label.svelte";
+export { default as FieldControl } from "./field-control.svelte";
+export { default as FieldDescription } from "./field-description.svelte";
+export { default as FieldError } from "./field-error.svelte";